requirements.txt
fastapi
uvicorn
gunicorn
supabase
requests
fastembed
python-dotenv
pydantic
pydantic[email]
pyjwt
passlib[bcrypt]
python-multipart
psutil
